# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Eventful API",
    "description": "**API documentation for the Eventful - An event managment application.**",
    "termsOfService": "",
    "contact": {
      "name": "Liasu Aanuoluwapo",
      "email": "liasu.olabayo@gmail.com",
      "url": "https://github.com/holabayor"
    },
    "license": {
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html",
      "name": "Apache 2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3300",
      "description": "Development Server"
    },
    {
      "url": "https://eventful-api-nxu6.onrender.com",
      "description": "Staging Server"
    }
  ],
  "paths": {
...
    "/events/categories": {
      "get": {
        "tags": ["Event Category"],
        "summary": "Retrieves all categories",
        "description": "Retrieves all categories",
        "operationId": "get-events-categories",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Get-All-Events"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "servers": [
          {
            "url": "http://localhost:3300",
            "description": "Development Server"
          },
          {
            "url": "https://eventful-api-nxu6.onrender.com",
            "description": "Staging Server"
          }
        ]
      },
      "post": {
        "tags": ["Event Category"],
        "summary": "Add a new category (SUPER ADMIN)",
        "description": "Add a new event category (SUPER ADMIN)",
        "operationId": "add-events-categories",
        "responses": {
          "201": {
            "$ref": "#/components/responses/200-Category"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "servers": [
          {
            "url": "http://localhost:3300",
            "description": "Development Server"
          },
          {
            "url": "https://eventful-api-nxu6.onrender.com",
            "description": "Staging Server"
          }
        ]
      }
    },
    "/events/categories/{id}": {
      "get": {
        "tags": ["Event Category"],
        "summary": "Get a category by ID",
        "description": "Get a category by ID",
        "operationId": "get-events-categories-id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the category",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200-Category"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "servers": [
          {
            "url": "http://localhost:3300",
            "description": "Development Server"
          },
          {
            "url": "https://eventful-api-nxu6.onrender.com",
            "description": "Staging Server"
          }
        ]
      },
      "patch": {
        "tags": ["Event Category"],
        "summary": "Update a category by ID",
        "description": "Update a category by ID",
        "operationId": "update-events-categories-id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the category",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200-Category"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "servers": [
          {
            "url": "http://localhost:3300",
            "description": "Development Server"
          },
          {
            "url": "https://eventful-api-nxu6.onrender.com",
            "description": "Staging Server"
          }
        ]
      },
      "delete": {
        "tags": ["Event Category"],
        "summary": "Delete a category by ID",
        "description": "Delete a category by ID",
        "operationId": "delete-events-categories-id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the category",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200-Category"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "servers": [
          {
            "url": "http://localhost:3300",
            "description": "Development Server"
          },
          {
            "url": "https://eventful-api-nxu6.onrender.com",
            "description": "Staging Server"
          }
        ]
      }
    }
  },
...
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "responses": {
      "200-Category": {
        "description": "Example response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "category": {
                  "$ref": "#/components/schemas/Category"
                }
              },
              "x-examples": {
                "Example 1": {
                  "message": "Category retrieval successful",
                  "category": {
                    "_id": "66b61949cfe178478521356f",
                    "name": "Business & Professional"
                  }
                }
              }
            }
          }
        }
      },
...
      "Get-All-Events": {
        "description": "Example response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "data": {
                  "$ref": "#/components/schemas/Category"
                }
              },
              "x-examples": {
                "Example 1": {
                  "message": "Event retrieval successful",
                  "data": {
                    "events": [
                      {
                        "_id": "66a24b50a1b78c41aa51165f",
                        "title": "Sapa Events 5.0",
                        "description": "Organised by Brown",
                        "date": "2024-07-26",
                        "time": "8:00AM",
                        "location": "Nigeria",
                        "creator": {
                          "name": "James Brown"
                        },
                        "attendees": ["66a636f89ea1d6315545aacf"],
                        "defaultReminderDate": "2024-07-25T07:00:00.000Z"
                      },
                      {
                        "_id": "66a3e23e4b7d10964cafb4b9",
                        "title": "Sapa Events 2.0",
                        "description": "Organised by Shetimma",
                        "date": "2024-07-28",
                        "time": "8:00AM",
                        "location": "Nigeria",
                        "creator": {
                          "name": "Scott Allen"
                        },
                        "attendees": ["66a636f89ea1d6315545aacf"],
                        "defaultReminderDate": "2024-07-27T07:00:00.000Z",
                        "eventQrCode": "https://res.cloudinary.com/dev-storage/image/upload/v1723111057/event_qr_codes/zntowjaxtebhblonltif.png"
                      },
                      {
                        "_id": "66ad245b40fd4671dd4786d7",
                        "title": "For Food Only",
                        "description": "Come and eat till your belle full",
                        "date": "2024-08-03",
                        "time": "8:00AM",
                        "location": "Lagos",
                        "creator": {
                          "name": "Sola Adeleke"
                        },
                        "attendees": ["66a636f89ea1d6315545aacf"],
                        "defaultReminderDate": "2024-08-02T07:00:00.000Z",
                        "eventQrCode": "https://res.cloudinary.com/dev-storage/image/upload/v1723111057/event_qr_codes/zntowjaxtebhblonltif.png"
                      },
                      {
                        "_id": "66af6bb257ff19ed664def9e",
                        "title": "The August Protests 3.0",
                        "description": "Oragnised by Thiefnubu",
                        "date": "2024-08-08",
                        "time": "8:00AM",
                        "location": "Nigeria",
                        "creator": {
                          "name": "Chigozie Alex"
                        },
                        "attendees": [],
                        "defaultReminderDate": "2024-08-07T07:00:00.000Z",
                        "eventDateTime": "2024-08-08T07:00:00.000Z",
                        "eventQrCode": "https://res.cloudinary.com/dev-storage/image/upload/v1723111057/event_qr_codes/zntowjaxtebhblonltif.png"
                      },
                      {
                        "_id": "66af6c1757ff19ed664defa2",
                        "title": "Osun Twitta Hangout",
                        "description": "Oragnised by Inside Osogbo",
                        "date": "2024-08-08",
                        "time": "18:00",
                        "location": "Osogbo, Osun State",
                        "creator": {
                          "name": "Hamid Muhammad"
                        },
                        "attendees": [],
                        "defaultReminderDate": "2024-08-07T17:00:00.000Z",
                        "eventDateTime": "2024-08-08T17:00:00.000Z",
                        "eventQrCode": "https://res.cloudinary.com/dev-storage/image/upload/v1723111057/event_qr_codes/zntowjaxtebhblonltif.png"
                      },
                      {
                        "_id": "66af6f1d57ff19ed664defa6",
                        "title": "Food Hangout 3.0",
                        "description": "Oragnised by Cleo",
                        "date": "2024-08-08",
                        "time": "08:00",
                        "location": "Nigeria",
                        "creator": {
                          "name": "Oluwaseun Cleopatra"
                        },
                        "attendees": ["66a636f89ea1d6315545aacf"],
                        "defaultReminderDate": "2024-08-03T17:00:00.000Z",
                        "eventDateTime": "2024-08-04T17:00:00.000Z",
                        "eventQrCode": "https://res.cloudinary.com/dev-storage/image/upload/v1723111057/event_qr_codes/zntowjaxtebhblonltif.png"
                      },
                      {
                        "_id": "66b226dfba82a6dc96fabf63",
                        "title": "Kano Hangout 3.0",
                        "description": "Oragnised by Hammid",
                        "date": "2024-08-08",
                        "time": "18:00",
                        "location": "Nigeria",
                        "creator": {
                          "name": "Hamid Muhammad"
                        },
                        "attendees": [],
                        "defaultReminderDate": "2024-08-07T18:00:00.000Z",
                        "eventDateTime": "2024-08-08T18:00:00.000Z",
                        "eventQrCode": "https://res.cloudinary.com/dev-storage/image/upload/v1723111057/event_qr_codes/zntowjaxtebhblonltif.png"
                      }
                    ],
                    "metadata": {
                      "page": 1,
                      "limit": 10,
                      "totalPages": 1,
                      "totalCount": 7,
                      "hasPreviousPage": false,
                      "hasNextPage": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "Category": {
        "type": "object",
        "description": "Event category",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "x-examples": {
          "Example 1": {
            "_id": "66b61949cfe178478521356f",
            "name": "Business & Professional"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}